+++ changelog.md
@@ -1,2 +1,8 @@
 # 1.0.0
 - hello geode world
+
+# 1.0.1
+- **Set-Profile-Music** added button on LevelInfo
+- **floating notes** added
+- *bouncing text*
+- Stopped supporting android :(